using regex to validate item code

+++ src/main/java/lk/ijse/possystembackendwithspring/controller/ItemController.java
@@ -1,6 +1,7 @@
 package lk.ijse.possystembackendwithspring.controller;
 
 import lk.ijse.possystembackendwithspring.dto.impl.ItemDto;
+import lk.ijse.possystembackendwithspring.exception.ItemNotFoundException;
 import lk.ijse.possystembackendwithspring.service.ItemService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -9,6 +10,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.regex.Pattern;
 
 @CrossOrigin(origins = "http://localhost:63342")
 @RestController
@@ -23,12 +25,20 @@ public class ItemController {
                                                @RequestPart("_itemName")String itemName,
                                                @RequestPart("_itemQty") String itemQty){
         try {
+            int qty;
+            double price;
+            try {
+                qty = Integer.valueOf(itemQty);
+                price = Double.valueOf(unitPrice);
+            } catch (Exception e) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
             ItemDto itemDto = new ItemDto();
             itemDto.setItemName(itemName);
             itemDto.setItemCode(itemCode);
-            itemDto.setItemQty(Integer.valueOf(itemQty));
+            itemDto.setItemQty(qty);
             itemDto.setCategory(category);
-            itemDto.setUnitPrice(Double.valueOf(unitPrice));
+            itemDto.setUnitPrice(price);
             System.out.println(itemDto);
             itemService.saveItem(itemDto);
             return new ResponseEntity<>(HttpStatus.CREATED);
@@ -42,18 +52,21 @@ public class ItemController {
 
         return "Success";
     }
+
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public List<ItemDto> getItemList(){
-        return itemService.getItemList();
+    public List<ItemDto> getItemList() {
+        List<ItemDto> itemList = itemService.getItemList();
+        return itemList;
 
 
     }
+
     @PutMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public void updateItem(@RequestPart("_itemCode") String itemCode,
-                           @RequestPart("_unitPrice") String unitPrice,
-                           @RequestPart("_category")String category,
-                           @RequestPart("_itemName")String itemName,
-                           @RequestPart("_itemQty") String itemQty){
+    public ResponseEntity<Object> updateItem(@RequestPart("_itemCode") String itemCode,
+                                             @RequestPart("_unitPrice") String unitPrice,
+                                             @RequestPart("_category") String category,
+                                             @RequestPart("_itemName") String itemName,
+                                             @RequestPart("_itemQty") String itemQty) {
 
         ItemDto itemDto = new ItemDto();
         itemDto.setItemName(itemName);
@@ -62,13 +75,47 @@ public class ItemController {
         itemDto.setCategory(category);
         itemDto.setUnitPrice(Double.valueOf(unitPrice));
         System.out.println(itemDto);
-        itemService.updateItem(itemCode,itemDto);
 
 
+        String regexForUserID = "^Item[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$";
+        Pattern regexPattern = Pattern.compile(regexForUserID);
+        var regexMatcher = regexPattern.matcher(itemCode);
+        try {
+            if (!regexMatcher.matches() || itemDto == null) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
+            itemService.updateItem(itemCode, itemDto);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } catch (ItemNotFoundException e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
 
     }
+
     @DeleteMapping(value = "/{itemCode}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public void deleteItem(@PathVariable("itemCode") String itemCode){
-        itemService.deleteItem(itemCode);
+    public ResponseEntity<Object> deleteItem(@PathVariable("itemCode") String itemCode) {
+
+        String regexForUserID = "^Item[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$";
+        Pattern regexPattern = Pattern.compile(regexForUserID);
+        var regexMatcher = regexPattern.matcher(itemCode);
+
+        try {
+            if (!regexMatcher.matches()) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
+            itemService.deleteItem(itemCode);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } catch (ItemNotFoundException e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
     }
 }
